Refactor TOML helper functions into `toml.rb`
Move TOML helper functions from `storage_volume.rb` to `toml.rb` so they
can be reused by other files.
